requirements.txt
streamlit>=1.0.0
scikit-learn>=1.0.0
nltk>=3.6.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=1.3.0
numpy>=1.21.0
